测试

+++ models/bot.go
@@ -746,11 +746,11 @@ var handleMessage = func(msgs ...interface{}) interface{} {
 						env := strings.Split(split[i], "=")
 						if !sender.IsAdmin {
 							coin := GetCoin(sender.UserID)
-							if coin < float64(Config.Tyt) {
+							if coin < Config.Tyt {
 								return fmt.Sprintf("推一推需要%d个积分", Config.Tyt)
 							}
-							RemCoin(sender.UserID, float64(Config.Tyt))
-							sender.Reply(fmt.Sprintf("推一推即将开始，已扣除%d个积分，剩余%f", Config.Tyt, GetCoin(sender.UserID)))
+							RemCoin(sender.UserID, Config.Tyt)
+							sender.Reply(fmt.Sprintf("推一推即将开始，已扣除%d个积分，剩余%d", Config.Tyt, GetCoin(sender.UserID)))
 						} else {
 							sender.Reply(fmt.Sprintf("推一推即将开始，已扣除%d个积分，管理员通道", Config.Tyt))
 						}
